Feature: add content to page
Used a module to contain functions that create elements and add them to
the page. Export it to index.js.

Added rule to webpack config file to handle image files.

Added div with id to keep all the page content. Created a css file.
Added image to use as a background.